Replace credhub path with variable name in bind
[#157735953]

At the moment the secrets map generated by ResolveManifestSecrets has a
map of credhub paths to bosh variable secret values. However, we were
concerned the service author would not know what the path was in order
to extract the secret value at bind time. So we removed the credhub
path, leaving only the variable name as the key to the secret.

This is one solution to resolve the secret value at bind time. We
discussed another solution where we simply allow the service author to
discover the credhub path by retrieving the path value out of the
manifest. If we decide that this is the better solution, then let's
revert this commit.

Co-authored-by: Derik Evangelista <devangelista@pivotal.io>